Postgres-related bug fixes:
561	PostgreSQL	**Create all tables without OIDs
The ALTER DATABASE statement sets the default for future sessions without affecting the current session. Fix is to also use SET to turn of OIDS for the current session.

563	PostgreSQL	**Global groups should always appear at the top
We were already sorting by ProjectId, but Postgres and SQLServer disagree on how to sort null, so we need to handle null ProjectId values as the first ORDER BY criteria.

The "basic" Verify test now passes on Postgres.