Massive changes to move us from multiple databases to a single database + schemas.  The changes are not complete and there are no doubt  some problems in less obvious areas.  Test your stuff carefully!
I've rewritten all queries in the issues, ms2, prot, and flow modules to be schema-independent.  Doing this allowed me to move these modules to new schemas on PostgreSQL but keep them in their old database + dbo on sql server.  Once all the code is moved and the migration scripts are written, we'll move everything on sql server as well.