Flow mini pipeline
The command line flow tool (/flow/build.xml) is now built as part of the web app build.  The jars get copied from /flow/build/dist/ into the deployed WEB-INF/lib directory.

You can now use the pipeline web part to find a directory containing .fcs files, and invoke the pipeline on it (i.e. generate some graphs).
You can also check the status of the pipeline (there's a queue of Runnable objects, only one of which is running at a time, and you can look at the status file of a running job, or look at the results of a completed job, or cancel a pending job).