Bug fix:
630	Build	Eliminate beehive/struts errors at startup

Beehive was cranky because there wasn't a config file for a controller at the root of the webapp. I added a no-op controller to get it to shut up.

The other "error" messages are really informational messages that are part of Beehive's error handling. There's nothing wrong, really, and to get rid of them we'd have to register an exception handler for every exception class that could ever be thrown in the entire app. It outputs a message as it walks up the class hierarchy until it does find a handler, which we have on Exception.class.

DRTs: passed