Bug fixes:
312 add attachment blows away pending edits (on IE at least)
379 Display text doesn't appear when wiki page hasn't been created
577 Minor problems with CPAS configured at root context path
632 Send email in group.view should not say fhcrc.org -- should use user domain
644 Can't manually recover verification email in case of send failure
661 Error is reported when adding existing user to a group
679 Success messages on adding users are funky
689 NavTrail Home shouldn't redirect
Also some minor installer fixes to get up-to-date with the latest CPAS.xml replacements.

All are pretty straightforward, with the exception of the wiki changes, which are slightly more complicated.  I've split editing the content and managing attachments into two separate pages to avoid the problems with data loss that can occur if content changes are made prior to an auto-refresh.  This whole UI can be reexamined after the 1.0 release- some cleaning up is needed.

All verify tests passed.