New tabbed navigation between modules.
Users can select which modules should be active in a given folder.  "Active" modules are displayed full-time as tabs at the top of the screen.  If the user ends up in a module that is not active in its folder, a tab is added.  Initial tab set is automatically calculated based on the webparts present in each folder's portal page.

Verify tests passed.