800	Bootstrap	Move app settings out of cpas.xml into database properties where it makes sense.
Values are imported from existing CPAS.xml if they're present in that file, and are set to defaults if not present.

Removed all caching of AppProps values that I could find. AppProps will cache for you, so always get a fresh value to ensure it hasn't changed since the last time you used it.

Changed Verify to require a username and password, which it uses to log in to the system. This gets rid of the Verify backdoor on dev builds.

DRT: passed