Create folder "/shared" for java files that are shared between flow.jar and the webapps. Currently, I just moved Cache into there, but we should move all of the non web-specific stuff from cpaslibrary into there.
Also, change Flow so that the step of finding FCS files on disk and verifying their keywords is separate from performing the analysis on them.
Note: this might not build right at this exact moment because I am having trouble with SVN, and am going to delete some folders in a subsequent checkin in a few minutes.