The upgrade of the properties table (exp-1.00-1.02.sql) was still failing when applied to the production snapshot 1.0 database on staging.  The problem was orphaned records in which the container field was null or was no longer a valid container.
Solution was to run a "delete orphans" script before running the upgrade, and also to prevent part of the problem by making container fields in the exp schema not null.  1.00-1.03 deletes orphans and makes the containers not null;  1.03-1.04 does the property table change.  For those that had already run the property table change, 1.02-1.04 deletes orphans and makes the containers not null (slightly different from 1.00-1.03 because of the table changes made  by the now-deleted 1.00-1.02)