Experiment cleanup work:
Get rid of BioSource and Fraction sub-types
   (including the template ms2 xar.xml that uses biosource.) these should just be Material objects. drop the BioSource and Fraction tables and move their contents into Property objects if necessary.

Get rid of handleColumnSeparation demo code that calculates fraction %buffer values based

Fix verify test that expects the above to work.

Add minOccurs=0 to MaxInput* and Output* elements of the ProtocolBaseType. If the element is not present, the behaior should be the same a nil=true, that is no generation of protocolapp instances nor output instances. also get rid of default=0 on these elements since it doesn't seem to do anything



(XarReader) Allow ProtocolDefinitions section to be completely empty. Currently you have to have the parent protocol listed in the Xar, but you could get the LSID of parent protocol from the ExperimentRun.ProtocolLSID element, then query the database to get all of the childprotocols, protocolActions and protocolActionPredecessors.

(XarReader) Allow ProtocolActionDefinitions to be completely empty.(currently need to have 1 entry to get it to work).

If ProtocolActionDefinitions is not empty but the parentProtocol already exists in the database, compare the databases' protocol actions with those in the xar and fail if they don't match in child protocols, action sequence numbers, and predecessors.

get rid of reference to gettypes in rdftypes schema

get rid of minOccurs=0 on all Name elements where it currently appears (not sure why I set to 0, causes problems in UI)

Constrain cpastype to one of the following:
     Material, Data, Protocol, ProtocolApplication, Experiment, ExperimentRun, ExperimentRunOutput via a pattern restriction

DRT: passed