Make JSPs compile faster. JSPs now get compiled into WEB-INF/jsp/modulename_jsp.jar. This enables you to recompile a JSP without bouncing the web app. Also, I made it so that doing a build when nothing has changed no longer causes the webapp to bounce. The webapp bounce had been caused by XML schemas, "cpas.xml", "verify.xml", and "modules.properties"
Also, I removed "preserveLastModified" from many of the <copy> tasks in the build script.
"preserveLastModified" should only be necessary if "overwrite" is "true".
This is because ANT normally does not copy the file if the target is newer, so it should not make a difference whether we set the lastModified to the current time, or the time on the file.