CruiseControl has been running fine for a few days; I'm now adding cpas-developers to the email-on-failure list.
This means that everyone will receive an email in the case of a build/DRT failure, even those who didn't check in and couldn't have caused the problem.  This is useful for notifying the group that it's not a good time to sync, and serves as a James-Corrigan-esque reminder of where the fingers should point.