223: Fix 'Customize Wiki Web Part' page
The customize wiki web part page now shows a drop-down list of containers and a second list of their pages, updated when the container is changed. We now store GUIDs for the container value, and page name for the page value. If the container does not have a wiki, a default page is listed in the drop-down, and selecting this page shows this page with no content.