1124 : Names generated into search.xars get too long
Fixed problem exporting protocols when two protocols have action steps that refer to the same protocols.

Changed relativized LSID template for Data, Material, and Sample objects so they will match up within an export XAR and not collide with other LSIDs that might already be on the system.

DRT: passed