Additional peptide-level data is now available for runs with ProteinProphet information.
Column picker for MS2 now includes ProteinProphet and Quantitiation columns if they are present for the run.

Protein views now show the peptides that ProteinProphet assigned to the protein.

Perf improvement to loading ProteinProphet info - we now load into a temporary table and move it to the permanent table, allowing us to do a single join to the sequences tables instead of many.

DRT: passed