added consolidated script for 1.1 to 1.2, and 1.2 to current 1.26.
added fk constraints between exp.Data, Material, and Object tables and the Containers table to find out who is orphaning data objects

removed demo schema test from hardcoded list, pending a better mechanism for asking each module to list what schemas they want to test.

screen out table and view names beginning with an _ from the xml consistency check.  If you have utility views that are not part of cpas but are in your dev databsae, they must either start with an underscore or be in a schema such as dbo that is not part of the checkin test.