First step in improving module upgrade.
SQLScripts are now run in a background thread. The user sees a page that lists the scripts being run for the current module upgrade. The browser uses AJAX to figure out when the script is done being run (reloading the page manually also checks), and bounces the user to the next step in the upgrade when it's done.

DRT: passed (including bootstrap)