Improved error handling:
If we get an exception during page rendering after the page has been partially written back to the client, we now try to dump the stack trace to the page, not just to the logs.

We now set status codes like 404 and 500 if we get an error rendering a page, assuming the response has not been committed. This lets us use automated link checkers against the site to see if all the linked pages are working correctly.

We now log all ERROR and FATAL level site-level log4j output to a separate file, in addition to cpas.log. The DRT checks to make sure there haven't been any errors logged before declaring a test a success. Additionally, the Admin Console lets you see all site errors since the server started, or since you last reset the errors.

Also, to pass the DRT with the error checks I had to fix:

1324 : WikiManager.purgeContainer() orphans wiki pages

DRT: passed