Upgrading to XmlBeans 2.1.0, to fix a classloader issue affecting msInspect when we run from viewerApp.jar.
Small code change required in Stax2DomBuilder:  in 2.1.0, XmlBeans
implements some slightly incorrect checking to make sure you're not
adding more than one child node to a document node.  This check actually
doesn't allow you to add _any_ nodes to a document node, so we have to
add a level of indirection when we build our node hierarchy.

Note: this XmlBeans upgrade will only occur on msInspect, for now,
because CPAS is still referencing the XmlBeans jar from
external/lib/web.  That should be changed whenever convenient!  The code
fix that addresses the 2.1.0 issue, however, WILL be pulled into CPAS by
this checkin.  The code fix will work on either 2.0.0 or 2.1.0, so this
is fine.

Passed DRT, postgres