First pass at the Mothership module. It currently collects exceptions and stores them for analysis.
In the Admin Console, you can configure the level of exception logging. You can turn it off, or choose from three different levels of reporting, depending on how comfortable you are with submitting potentially private data (like the URL or email address of the user).

There's a fairly simple UI for browsing and dealing with the submitted exceptions. You can assign a bug number or a particular person who should investigate.

If the -Ddevmode=true or -Dgroovy.root system properties are set, exceptions will be logged to the local server. If not, they'll be logged to cpas.fhcrc.org.

DRT: passed