Usage reporting and upgrade check
By default, CPAS installs now phone back to the mothership and report that they're alive every 24 hours. This can be turned off on the Customize Site page. There are two levels of reporting detail, which are admin-configurable.

If an installation is contacting the mothership, it also checks if a new version is available. If so, it displays a message in the page header for admins.

Installing or upgrading now dumps the user in the Customize Site page at the end of the process, where they can set the reporting level.

DRT: passed (Postgres and SQLServer)