Enhanced CommandLineModuleDiscoverer to work with the one-jar viewerApp deployment.  The one-jar deployment presents, shall we say, interesting challenges for discovering extensions.  Here's the deal:
one-jar has a Very Special classloader called JarClassLoader.  It's Very
Special in that you can't interrogate it to find the jar files that the classes are coming from, as resources.
You can only locate them as entries on the main jar itself.
So, if we're using that classloader, we have to follow a very
special convention, which is this:
     -Each extension jar contains exactly one extension
     -All extension jars go in the lib directory of the one jar
     -Each extension jar has the name "extension_ClassName.jar",
      where ClassName is the name of the CommandLineModule class contained in the jar

Fun, eh?  This will have to do until we're forced to spend a lot of time
on something better.

Also slightly better help messages, and slightly more usage info in
AmtCommandLineModule.

viewer/workbench only