AMT changes to incorporate the idea that each run has two regression lines that are important:  one determines hydrophobicity as a function of time, the other determines time as a function of hydrophobicity.
In addition, I discovered some problems with BasePepXmlWriter.  It was
writing out the wrong index for aminoacid modifictions (should be
one-based, it was writing out 0-based), and it now also writes
modification mass as well as massDiff.


viewer-only, except for the changes to BasePepXmlWriter, which are
minor and involve no dependencies; no one uses BasePepXmlWriter but me.