Logo image tweaks:
1400: admin: file name for custom logo, icon should be displayed in field

Also, don't require read permissions on the root container to get the logo (most database instances allow this, but Proteomics does not).

Additionally, add a counter to the logo <img> tag so that we don't allow the user's browser to keep pulling an outdated image from its cache.

DRT: passed