Major changes to both the CommandLineModule structure and FeatureGrouper.
FeatureGrouper now uses the new class FeatureClusterer under the hood,
which in turn uses the generic Clusterer2D.  This separates the actual
work of clustering from the rest of the peptide-array creation stuff
that FeatureGrouper does.  The behavior is also modified somewhat:
instead of summing intensities in the peptide array, we pick the feature
closest to the middle of the bucket and use that feature's intensity to
represent that FeatureSet in the bucket.  Also, we carry through the
identified peptide and protein of the feature into the peptide array, if
available.

As for CommandLineModule, I drastically changed the way arguments are
declared and handled (especially declared).  No backward compatibility
on argument declaration.  Check FilterFeaturesCommandLineModule or
the new PeptideArrayCommandLineModule for good examples.

Passed msInspect DRT.  Checkin is viewer/workbench only.