Somewhat major msInspect refactor.  All non-core fields in viewer.feature.Feature have been stuffed into a Map in Feature, rather than stored in defined fields.  This gives us the ability to define flexible sets of fields of any type that can be added to Feature.  These sets of fields are described in feature.extraInfo.FeatureExtraInformationDef and its subclasses.  The initial sets are: MS2, IsotopicLabel, AMT, CID, Time. For more information, see the extensive comments in FeatureExtraInformationDef.
Cascading changes throughout the code to deal with this, plus a minor
change to TabLoader, to allow it to call put() on classes that implement
Map.  

Also added a "best" resolver to resolve conflicts between features from
the same featureset in the same bucket in peptide arrays.  "sum" is
still the default.

viewer-only, except the TabLoader change.