Partial checkin due to subversion weirdness.  More files to come shortly. Apologies if build is temporarily broken.
Creation of "ExperimentService" interface.  This interface is the public API for the experiment module.  It's still a work in progress.
A notable missing part of the ExperimentService is OntologyManager-- I hope to give that into a public API soon.
Refactoring of Flow module to use ExperimentService
Change to XarReader to set the OwnerObjectLSID of data's etc. to be the data object's LSID instead of the Run's.
Modify uploading of SampleSets.  The user can now choose up to 3 key columns.
Also, the user can upload additional samples to an existing SampleSet.
Change FlowAnalysisScript to be a Data (it used to be a protocol).  The Flow protocol is now the place to put properties about the protocol, such as how to link FCS files to samples.