Re-enable caBIG DRT after tweak to allow test harness to delete pipeline jobs successfully.
Allow multiple experiment sections in a XAR. If there are multiple entries, experiment runs must specify an ExperimentLSID. If there is just one experiment in the file this is optional. In the future, when we support many-to-many experiment/run mappings, we can allow multiple ExperimentLSIDs on the runs.

DRT: passed