Aggregates (SUM, AVG, MIN, MAX, COUNT) in data regions.
Data regions can now display aggregate query information in a row at the top and/or bottom of the region.  To use, call setAggregates on the region, specifying which columns and which functions.  The default placement of the information is at the bottom of the data region, but this can be altered by calling setAggregateRowPosition.  Note that the current implementation essentially performs the data region's query twice, once for the rows and then again for the aggregates- this was the simplest implementation, though I'm open to any brilliant approach that improves performance.

This functionality was required for CHAVI issue 1638, "Specimen view availability summary".

New functionality tested on SQL Server and PostgreSQL.  DRT passed on SQL Server.