Make all result objects immutable.
Change API to make it much clearer that you need to set a context to make a query (parentPath for folders, folderPath for protocols and experiments, and experimentLSID for runs).

Add tests for immutability and for getting the content of files.

JavaDoc the methods that don't correspond to UML elements, add caBIG JavaDoc target to build.

DRT: passed