1681 : Missing protocol for samples loaded after fraction run
It wasn't showing the protocol and sample because we skipped the code path that populated the relevant information when we thought that it was just an intermediate file on the way to generating a combined all.pep.xml. We now populate the information in the other possible code path as well.

DRT: passed