1962 : Change ProteinProphet percent storage in database 2035 : Deadlock: ModuleClassLoader.loadClass vs. WebAppClassLoader.findClassInternal 2032 : Review usages of PipelineService.getPipelineRoot() 2045 : User authentication for users of remote client api
Also, reduce caBIG devkit to include just the necessary JARs.

Add a Jump to issue box for issues. Convert buttons to links, per the newest version of the spec.

Add SiteDownServlet to web.xml so that it's easy to enable. Add init-param to allow for customizing the message.

DRT: passed