1423: **Add Change Notification field to Issue Tracker
The change notification field does not get populated or modified automatically although there is a new email preferences link off the grid view to allow modification of the email notification rules: 
-Send me email when an issue is opened and assigned to me
-Send me email when an issue that's assigned to me is modified
-Send me email when an issue I opened is modified
-Send me email notifications when I enter/edit an issue

The current default is for the first 3 selections to be on. Also, refactored some of the issue views to be jsp-based.


507: Expand phone numbers to 64 characters

wasn't checking the length after the extra characters were being added, added friendlier messages when the length is exceeded.

drt: passed