Full rebuild required after syncing to this checkin.
Changes to let IntelliJ HotSwap classes while debugging. More details to follow in a separate email.

Fix for ThreadDeathError when redeploying the webapp - updated to commons-logging 1.1, which contains fixes for removing loggers in application servers so that they don't try to write after the webapp is shut down.

Start for new Experiment Run view, a replacement for the MS2 Runs view and the Experiment Navigator. Far from finished.

2080 : Send jump-to-issue form to a validation action

DRT: passed