Adding the 'interactive' command-line command for msInspect.
When invoked on a particular command, pops up a dialog box that allows
the user to specify all the commandline arguments in separate text
fields, and indicates which ones are required.  When they click Execute,
it runs the command.

This was low-hanging fruit... it may or may not have been worth the
effort, but it sure wasn't much effort.  If it ends up being used a lot,
it could be made much more usable by using different types of components
(e.g., dropdowns) for different types of arguments, and by providing
mouseover or tip text describing valid argument values.

Oh, one more thing, it saves previously-specified values for the command
to Preferences.  So you can save your preferred parameters for a command
without having to cut-paste the command to a file or something.  That's the bit that, I think, makes it worthwhile.