New plate service, and rewrite of NAB to utilize the service.
NAB behavior should be unchanged from before, but all processing is now driven by a plate template/layout that is applied to the plate data.  The NAB template is currently hard-coded, but changes in 1.8 should allow users to define new plate layouts and apply NAB-like processing to them with minimal effort.

Old NAB data should be upgraded on first access of the container; this will need further testing before the release of 1.7.

DRT passed on postgres.