2163: Support [Describe] and [Select] from MS2 Runs grid and Sample details page
Added link from Data details page to associated pipeline directory for the current container, if it exists and has been configured. This lets users describe a file if they haven't already.

Additionally, added info on the MS2 search protocol page showing the files and/or runs that are currently annotating the file. Added feature to let users redescribe, which will delete the existing runs and files and reprompt them to annotate.

Also, don't restrict materials in the run grids to the current container's active sample set.

DRT: passed