Improve deletion of samples and sample sets for 2033: Improve selection page for starting sample of Ms2 Run.
Check all materials in a sample set when deleting. If any are in use in experiment runs, show the user the full list to confirm that they should be deleted. If they are in different containers, call that out. If they are in containers where the user does not have permission to delete them, disallow the delete and tell the user why.

Same thing goes if individual samples within a set are being deleted.

Fix null protocol in run name when using a shared protocol for multiple .mzXML files.

DRT: passed