Publish to study from NAB.  This simple feature allows nab runs to be published to a study using some of the mechanisms already provided by the generic assay manager.  Published runs look just like a normal assay dataset, but contain a new source lsid column, which allows for references back to the originating data.
Also found the cause for our intermittent study DRT failure: the word 'complete' was found in one of our dataset names, causing the test to believe that the pipeline was done loading data before it was actually complete.  I've modified the test to fix this issue.

DRT passed on SQL Server and Postgres.