2000 BatchUpdateException creating a NAB run with sample data Negative and positive infinity values are now saved as max and min double and mapped back to infinity values on retrieval.
1429 site customization: from address may now overwrite address user specified in installer
We no longer ask for from address in the installer- this value was ignored anyway.

2214 Improve decimal formatting for NAB
Only three decimal places are now shown, rather than the previous 10-12.

2263 study: testing buddy finds TestStudy Study Study
I've removed some of the redundant "study" text on the dashboard page.

1919 Add a print view for the graph selected specimens view
Print view is now available from single and multi-run view.

2264 Publish fails when cutoffs are blank
Default cutoff values are now saved to the database when no values are entered.  Before, we would save to the blanks to the DB, and use in-memory default values at calculation time.  This caused several problems with data region display code that assumed that DB cutoff values would never be null.  

2125 Add ability to delete unused actors and statuses from specimen request setup
Simple delete links have been added.  These links are only available for statuses and actors that are not currently in use.

2265 Cutoffs can be nonsensical values
Cutoffs are now restricted to values between -100 and 200.  (Below 0 and above 100 are actually appropriate percentages for NAB.)

2322 Non-curve-based cutoff values should be added to NAB results view
2321:NAB: Provide more space for sample id
2324:NAB: FileID can be auto-populated from file name
These three simple tweaks came out of the NAB demonstration this morning; they'll greatly help David Montefiori's lab get output in the format they like.

DRT passed on both SQL Server and Postgres