Bug 1924:  Protect pages from being taken down by a web part crash.
This was a fairly involved bug, but most of the work was simple refactoring.  Changes are:
--Moved much of Global.java rendering of exceptions into ExceptionUtil.java (in api).  This file has methods to return a ErrorView (HttpView), an ErrorWebPartView (WebPartView), or an ErrorRenderer (inner class of ExceptionUtil).
ErrorRenderer actually renders the guts of the exception details (stack, relevant info from the context, etc...).  The two views render it in different ways (whole page = ErrorView, webpart = ErrorWebPartView).

--Made WebPartView.prepare, doStartTag and doEndTag final.  doStartTag and doEndTag were not being overridden, so this was just precautionary.  Then I added prepareWebPart method to WebPartView and refactored all derived classes prepare methods to be prepareWebPart methods.  This was safe because the super.prepare() was empty (HttpView.prepare is empty).

--The guts of the change were to put exception handling in WebPartView.prepare and WebPartView.renderView to catch the exceptions, generate an error renderer and render that instead of any other content.  This isn't perfect as you could get some broken rendering, but it prevents the webpart from taking down the page, and gives you the usual exception info in a compact package (it's in a scrolling div that is limited in size).

--Also made WebPartFactory call to the constructor of a web part wrapped.  If, the constructor throws an exception an ErrorWebPartView is substituted in place (with the title of the original WebPart).

The upshot is that there should be little to no changes, except that web parts will now render exceptions inside their space instead of being thrown out to the Global handler.

I tested all this by placing exceptions strategically in the DemoWebPart during testing.  All seemed to work.  Additionally, the normal exception handling (in Global.java) seems to be the same, but if anyone sees any anomolies or features of the exception rendering missing, please file a bug, as i likely broke it.

I'm also wondering if there are exceptions that should be let through (security violations?) to be handled in other ways.  If so, let me or britt know.

DRT: pass
Forced DemoWebPart exceptions