Bug 1930:  Search box on issues grid view.  This box runs the standard search, but limits it to only issues.  Still renders results the same way and doesn't restrict the search in some useful was (i.e. only open bugs, etc...), but it's a step in the right direction.
As part of this, i refactored the Search classes methods to support taking a list of modules to search.  The default behavior is still to search all searchable modules.  Also fixed up the 'getSearchResultNames()' method to format correctly for 1, 2 or 3 modules (although i only tested with 1 and 3).

Tests: Added issue module to local server and typed in several search queries.  Added search web part and verified that old search behavior is unchanged.
DRT: pass

:)ross