2326: NAB publish should define virus id as additional key field This was straightforward after receiving Matt's schema fixes.  We simply specify virusid as an additional key property.
2215: Perform NAB data migration as part of schema upgrade
I've moved the migration code into afterSchemaUpgrade.  Additionally, I verified that the upgrade succeeds on a copy of the current Atlas data.

1960: New Grid View should have own view name
This required a small change to ReportManager and the Report interface to give Report types the opportunity to perform additional cleanup on deletion.  Code reviewed by Karl.

DRT passed on Postgres.