Add "Date" (yyyy-MM-dd) and "DateTime" (yyyy-MM-dd HH:mm) named formats and replace all schema xml date formats.
Handle null in DateUtil.format*() and remove PageFlowUtil versions.  Add formatDate() and formatDateTime() to JspBase -- JSPs can just call these to format dates using the standard formats.

Add TemplateBean.setAnchor() -- set a named anchor and the page with automatically scroll to that location.  This approach preserves the anchor on login redirects and address problems with anchors in the DRT harness.

2497 : Links to named anchors on thread.view don't survive login redirect
2536 : Changes to date handling