2548: BVT failure: Intermittent assertion error in OntologyManager.importTypes due to non-thread safe assumption
OntologyManager.importTypes contains the following two lines of code (1813,1814 as of r5406):
     DomainDescriptor dd = (DomainDescriptor) domainDescCache.get(getCacheKey(dURI, container));
     assert (null!= dd);
This code does not synchronize domainDescCache, so it cannot assume that the cache hasn't been cleared by another thread. The correct behavior is to replace the first line with this:
     DomainDescriptor dd = getDomainDescriptor(dURI, container);
This will get the DomainDescriptor from the cache if available, and will re-retrieve it from the DB if not.

DRT and BVT passed on SQL Server.