Added core,fn_dropifexists (procedure on mssql, function on postgress) that will test for the existence of a table, view, index, or whole schema and drop them if the object exists, do nothing if the object doesn't exist.  You can use these directly in scripts as I do in ms2-1.70-1.71.sql where I needed to drop an index if it exists before defining it.   You can use the same functionality in code through the methods DbSchema.dropTableIfExists, dropIndexIfExists, which use the underlying SQLDialect method dropIfExists.  Also changed the SQLDialect method dropSchema to use fn_dropifexists which keeps it from throwing an exception if the schema does not exist.  ADded junit test for these functions.
changed PlateManager to not use the AttachmentService as a static so that bootstrapping with no cpas database works

Added 3 new indexes corresponding to foreign key columns in the MS2 tables.  Since I already added these indexes to proteomics I used the new fn_dropifexists to prevent errors when proteomics is upgraded.

Improved the DbChecker admin tool. now uses dropifexists so no exceptions show up in the log, plus I changed the orphaned container handling to offer to recover orphaned objects instead of spitting out DELETE statements for the user to run. Use the "click here" link at the end of the dbchecker container check messages.   It recovers orphaned objects by creating a new project named _RecoveredObjectsxxxx and attempting to move the orphaned objects into this project.  Sometimes this will fail because the orphaned objects came from two different containers and fail a uniqueness test in the recovered container.  Run the dbchecker again and hit the fix link and it will create another recoveredobjects project and attempt to move the remaining orphans to that project.  

Once the orphans are in a new project the standard UI can be used to investigate them and presumably delete them if no longer needed.