Study designer` for buddy tester to break. Here's what works: Should be able to add a StudyDesigner web part & create study designs. You should be able to save and come back later and get full fidelity.
You can add timepoints and gruops to the immunization schedule and click on the cells to define immunizations in terms of adjuvants and assays.

You should be able to add timepoints for assays then click to have them executed and click to on the label to define amounts of materials needed.

Known Bugs: If you don't have any assays at a timepoint it won't be preserved via save.
Deleting Assays has unknown consequences.