Cancel queries that take a long time when the user hits "stop" in the browser or navigates to another page, or otherwise terminates the request.
RenderContext has a method "setAsync" on it, which causes the query to be run on a separate thread.
Meanwhile, the main thread sends a space character every ten seconds to the response stream.
If the user has cancelled the request, response.flush() will throw an IOException.
If this happens, we call "cancel" on the java.sql.Statement, causing the query to get cancelled.