Combine the run and run data upload steps. Saves the user at least one page visit per run upload.
Don't prompt for sample ids when uploading a generic assay, since we don't actually use the values.

Make the ignore warnings a checkbox instead of a separate button.

Get rid of AssayRunType and instead move to separate actions, specified by the AssayProvider.

Use the provider to create the special link to go to a particular run's data.