3996 : The name of an assay, not a [details] link, should link to the assay's run list
Also, 7x perf improvement for Luminex upload - by default the Excel reader we're using loves to kick off System.gc(). I've disabled it so that it doesn't for all of the uses. Luminex gets the biggest speed bump since the reader was garbage collecting for each sheet, but others uses get approximately 25% faster as well.