4019: Wiki TOC no longer displays content from other folders
The logic to get the container to use for a particular web part doesn't work before render (or perhaps prepare webpart).  In the refactored code, i was generating the tree in the constructor for the TOC.  I've moved this to the render method, but it required adding setters for id and elements to NavTreeMenu.  Should be revisited in 2.3.  Probably TOC should 'include' a NavTreeMenu rather than deriving from one.

CR: brittp
:)ross