Partial fix for 4181: Improve NAb graphing capabilities
We now perform a five parameter fit, as well as some additional tweaks to try to detect when the beginning or ending data is not at a neutralization plateau.  Because the graphing parameters are rather complex, I've removed the settings that end-users were once able to tweak.  No one ever changed them before, and the new settings would be too confusing.

Still remaining for this bug is the ability to lock the axis of the graphs; this will have to wait until after I convert nab to use the generic assay infrastructure.