Selenium disables all browser caching, resulting in about 38,000 requests per DRT run.  This check-in introduces -Dcaching=true support to both Selenium server (enabling in memory browser caching only, still no disk) and LabKey server (enabling .js file caching), and turns it on by default when starting Selenium server for the DRT.  This browser caching of images alone reduces the requests during a DRT run to about 13,000.  If you also start Tomcat with -Dcaching=true, Tomcat default streaming for .js files will be used, and they will also get cached.
Using both yields a DRT run that makes only 3,200 requests.  Unfortunately not a big perf win on time, but might be worth using for CruiseControl.  Just trying to put the time I spent debugging the virus scanner throttling of my machine to some use.

Using -Dcaching=true on Tomcat/LabKey is not recommended, if you are actively working with .js files (including GWT).