1704: Generic Assays: Out of Range Support
NAB now supports out of range values automatically for its calculated IC/EC values.  Other assays can add a column suffix "OORIndicator" to a column in their data domain to enable the same functionality.  This is not documented or discoverable for now; we need significant improvements to the domain editor UI to enable some sort of "enable out of range" checkbox.

DRT and Assay BVT passed.