Fixed 4447 : Form validation is not hooked up (wiki)
Also fixed Wiki DRT as it was incorrectly assuming that you can create a wiki page with no body content. This used to be possible only because the validation code was not properly hooked-up after the conversion to Spring. Now that I've fixed that, the DRT began to fail, so I updated it to add some content to the page.