Fixed 4338 : Radeox renders malformed nested bullets and 4189 : Extra line added before second-level bullets in inserted wiki pages
Same fix resolved both issues. Radeox's ListFilter was written such that it always wrote <li>...</li> tags for each line, even if there were nested lists. I enlisted in the radeox SVN project, and changed the ListFilter so that it emits the </li> tag only if the current bullet is the same as the previous one, or when poping up a level. 

Tested against the repro cases in the bugs and with several pages from the labkey.org site that contain nested lists.

From what I can tell, we were running 1.0 beta 2. The current source has a .version file which reports 1.1 BETA. However, their changes.txt file reports that it is 1.0 BETA 3, and there are no change entries past BETA 2. So it's a bit unclear as to what version it really is. Their web site (radeox.org) has absolutely no info as to the version.

Therefore, if these new jars cause any noticeable problems with wiki rendering, we can revert this change and implement the fix as a local filter in our own code.